replaced coredata with swiftdata model
