+++ PokeDex/FetchService.swift
@@ -14,7 +14,7 @@ struct FetchService {
     
     let baseURL = URL(string: "https://pokeapi.co/api/v2/pokemon")!
     
-    func fetchPokemon(_ id: Int) async throws -> FetchedPokemon {
+    func fetchPokemon(_ id: Int) async throws -> Pokemon {
         let fetchURL = baseURL.appending(path: String(id))
         
         let (data, response) = try await URLSession.shared.data(from: fetchURL)
@@ -26,7 +26,7 @@ struct FetchService {
         let decoder = JSONDecoder()
         decoder.keyDecodingStrategy = .convertFromSnakeCase
         
-        let pokemon = try decoder.decode(FetchedPokemon.self, from: data)
+        let pokemon = try decoder.decode(Pokemon.self, from: data)
         print("Fetched pokemon: \(pokemon.id): \(pokemon.name.capitalized)")
         return pokemon
     }
