+++ PokeDex/Pokemon.swift
@@ -1,23 +1,31 @@
 //
-//  FetchedPokemon.swift
+//  Pokemon.swift
 //  PokeDex
 //
-//  Created by Vinod Rathod on 19/06/25.
+//  Created by Vinod Rathod on 27/06/25.
 //
-import Foundation
+//
 
-struct FetchedPokemon: Decodable {
-    let id: Int16
-    let name: String
-    let types: [String]
-    let hp: Int16
-    let attack: Int16
-    let defense: Int16
-    let specialAttack: Int16
-    let specialDefense: Int16
-    let speed: Int16
-    let spriteURL: URL
-    let shinyURL: URL
+import Foundation
+import SwiftData
+import SwiftUI
+
+@Model
+class Pokemon: Decodable {
+    @Attribute(.unique)var id: Int
+    var name: String
+    var types: [String]
+    var hp: Int
+    var attack: Int
+    var defense: Int
+    var specialAttack: Int
+    var specialDefence: Int
+    var speed: Int
+    var spriteURL: URL
+    var shinyURL: URL
+    var sprite: Data?
+    var shiny: Data?
+    var favorite: Bool = false
     
     enum CodingKeys: CodingKey {
         case id
@@ -44,10 +52,10 @@ struct FetchedPokemon: Decodable {
         }
     }
     
-    init(from decoder: any Decoder) throws {
+    required init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         
-        id = try container.decode(Int16.self, forKey: .id)
+        id = try container.decode(Int.self, forKey: .id)
         
         name = try container.decode(String.self, forKey: .name)
         
@@ -71,22 +79,73 @@ struct FetchedPokemon: Decodable {
         }
         types = decodedTypes
         
-        var decodedStats: [Int16] = []
+        var decodedStats: [Int] = []
         var statsContainer = try container.nestedUnkeyedContainer(forKey: .stats)
         while !statsContainer.isAtEnd {
             let statsDictionaryContainer = try statsContainer.nestedContainer(keyedBy: CodingKeys.StatDictionaryKeys.self)
-            let stats = try statsDictionaryContainer.decode(Int16.self, forKey: .baseStat)
+            let stats = try statsDictionaryContainer.decode(Int.self, forKey: .baseStat)
             decodedStats.append(stats)
         }
         hp = decodedStats[0]
         attack = decodedStats[1]
         defense = decodedStats[2]
         specialAttack = decodedStats[3]
-        specialDefense = decodedStats[4]
+        specialDefence = decodedStats[4]
         speed = decodedStats[5]
         
         let spriteContainer = try container.nestedContainer(keyedBy: CodingKeys.SpriteKeys.self, forKey: .sprites)
         spriteURL = try spriteContainer.decode(URL.self, forKey: .spriteURL)
         shinyURL = try spriteContainer.decode(URL.self, forKey: .shinyURL)
+    }
+    
+    var spriteImage: Image {
+        if let data = sprite, let image = UIImage(data: data) {
+            Image(uiImage: image)
+        } else {
+            Image(.bulbasaur)
+        }
+    }
+    
+    var shinyImage: Image {
+        if let data = shiny, let image = UIImage(data: data) {
+            Image(uiImage: image)
+        } else {
+            Image(.shinybulbasaur)
+        }
+    }
+    
+    var background: ImageResource {
+        switch types[0] {
+        case "rock", "ground", "steel", "fighting", "ghost", "dark", "psychic":
+                .rockgroundsteelfightingghostdarkpsychic
+        case "fire", "dragon": .firedragon
+        case "flying", "bug": .flyingbug
+        case "ice": .ice
+        case "water": .water
+        default: .normalgrasselectricpoisonfairy
+        }
+    }
+    
+    var typeColor: Color {
+        Color(types[0].capitalized)
+    }
+    
+    var stats: [Stat] {[
+        Stat(id: 1, name: "HP", value: hp),
+        Stat(id: 2, name: "Attack", value: attack),
+        Stat(id: 3, name: "Defense", value: defense),
+        Stat(id: 4, name: "Special Attack", value: specialAttack),
+        Stat(id: 5, name: "Special Defense", value: specialDefence),
+        Stat(id: 6, name: "Speed", value: speed)
+    ]}
+    
+    var highestStat: Stat {
+        stats.max {$0.value < $1.value }!
+    }
+    
+    struct Stat: Identifiable {
+        let id: Int
+        let name: String
+        let value: Int
     }
 }
